feat: implementa edicao de paciente com navegacao por etapas e melhoria de usabilidade (#519)
